Upgraded UI/Fixed Bugs
These changes in the UI of the CHeaderParser application reflect the addition of the DotNetCoolGridView and DotNetCoolButton controls from the DotNetCoolControls component library I created.  In addition, a bug in the StructExtractor class of the CHeaderParser.Extractor library was fixed, which was not extracting structure/union fields that had spaces in their element declarations.