Fixed Space in Numeric Array Field Declaration Bug
Fixed a bug in the StructExtractor class of the CHeaderParser.Extractor library and the ExtractParseUtils class of the CHeaderParser.Utils library.  The extractor algorithm was not extracting structure/union fields that had spaces in their element declarations.  I fixed the bug, by treating array declarations with spaces the same as other non-numeric numbers, which allows the non-numeric element conversion functions in the ExtractParseUtils class to resolve the spaces and convert the array element into the appropriate numeric value.