Upgraded UI/Fixed Bugs
These changes in the UI/Sample Application of the CHeaderParser application reflect the addition of the DotNetCoolGridView and DotNetCoolButton controls from the DotNetCoolControls component library I created.  In addition, a bug in the StructExtractor/ExtractParseUtils classes, which was not extracting structure/union fields that had spaces in their element declarations.