Added New Features to DotCoolButton
I added some new impressive features to the DotCoolButton control which is used in the CHeaderParser.UI project and application.  The new features include allowing different gradients and button caption text settings to be displayed in the button based on whether the button is in its normal state or whether the mouse cursor is positioned over the button or the mouse button is pushed down on the button.  This allows for some very nice visual effects, that allow for the button to be more interactive, such as gradient morphing, background and text lighting, bolding and color changes based on if the button is being hovered or clicked.